+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from './types';
-export * from './config';
 export * from './create';
 export * from './component';