+++ src/svg.ts
@@ -1,11 +1,10 @@
 import {CustomCreator, TaggedCreator} from './types';
 import {createElement, getTaggedCreator, getTaggedCreatorMap} from './create';
-import {defaultConfig} from './config';
 
 export const svgNamespaceUri = 'http://www.w3.org/2000/svg';
 
 export const s: CustomCreator<SVGElement> = (tagName, ...args) =>
-  createElement(svgNamespaceUri, tagName, defaultConfig, args) as any;
+  createElement(svgNamespaceUri, tagName, args) as any;
 
 type Result = {
   [K in keyof SVGElementTagNameMap]: TaggedCreator<SVGElementTagNameMap[K]>;
@@ -76,7 +75,7 @@ export const {
   use,
   view,
 } = getTaggedCreatorMap<Result, keyof SVGElementTagNameMap>(
-  getTaggedCreator(svgNamespaceUri, defaultConfig),
+  getTaggedCreator(svgNamespaceUri),
   [
     'a',
     'animate',